decoupled logic for the client, JSON and Connection entities
